Bug fixes from Michael Pöttgen:
improve [ 599719 ] "<see> tags in namespace summaries". It does not
remove the first two letters when a typename is passed as a cref, instead
of a type id. But there will be no link, unless you pass a valid type id.

improve [ 595560 ] "Detect "new" members". The additional tags for
hidden methods are not generated anymore. Also when a method hides
another method, the "new" keyword is added to c-sharp syntax. I havent
looked at other member types yet. Also I think there is a bug in vb-syntax
with the "Overloads" keyword.

fix [ 606305 ] "Overload of inherited method".